feat:SI_SDR, LMSE and SSIM metrics were added. Documentation complete
For more audio quality referencing, SI-SDR was added to show the drop in
audio quality. Laplacian Mean Squared Error (LMSE) and Strucutral
Similarity Index (SSIM) was also added as metrics for the watermark
preservation. A no_attack row was also generated in the output csv, so
that the effect of the extraction process could be quantified.